mmc: sdhci: changing sdhci quirks data type to u64
Kernel 3.1 has more than 32 quirks
changing quirks type to u64

Bug 921653

Reviewed-on: http://git-master/r/73167

+++ drivers/mmc/host/sdhci-pltfm.h
@@ -17,7 +17,7 @@
 
 struct sdhci_pltfm_data {
 	struct sdhci_ops *ops;
-	unsigned int quirks;
+	u64 quirks;
 };
 
 struct sdhci_pltfm_host {
